medical-ai-chatbot-1/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
openai==0.28.0
webrtcvad==2.0.10
pyaudio==0.2.11
python-dotenv==1.0.1
requests==2.32.3
sounddevice==0.5.1
soundfile==0.13.1
SpeechRecognition==3.14.2
numpy==2.1.3
pydantic==2.10.6
pydantic_core==2.27.2
gunicorn==20.1.0  # Explicit version to ensure it's installed
